.ipynb_checkpoints/requirements-checkpoint.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=gpu
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.6.3=py37he774522_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py37_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2020.12.8=haa95532_0
cachetools=4.2.0=pyhd3eb1b0_0
certifi=2020.12.5=py37haa95532_0
cffi=1.14.4=py37hcd4344a_0
chardet=3.0.4=py37haa95532_1003
click=7.1.2=py_0
colorama=0.4.4=pyhd3eb1b0_0
console_shortcut=0.1.1=4
cryptography=3.3.1=py37hcd4344a_0
cycler=0.10.0=pypi_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dnspython=2.0.0=pypi_0
entrypoints=0.3=py37_0
eventlet=0.24.1=pypi_0
gast=0.3.3=py_0
google-auth=1.24.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
greenlet=0.4.15=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.10=py_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.2=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.19.0=py37hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py37_0
jinja2=2.11.2=py_0
joblib=1.0.0=pypi_0
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py37haa95532_0
jupyterlab=2.2.6=py_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=1.2.0=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.3.1=pypi_0
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.3=py37haa95532_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.3.3=pypi_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2020.2=256
mkl-service=2.3.0=py37h196d8e1_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
monotonic=1.5=pypi_0
moviepy=1.0.3=pypi_0
msys2-conda-epoch=20160418=1
multidict=4.7.6=py37he774522_1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
nest-asyncio=1.4.3=pyhd3eb1b0_0
notebook=6.0.3=py37_0
numpy=1.18.5=pypi_0
numpy-base=1.19.2=py37ha3acd2a_0
oauthlib=3.1.0=py_0
opencv-python=4.4.0.46=pypi_0
openssl=1.1.1i=h2bbff1b_0
opt_einsum=3.1.0=py_0
packaging=20.8=pyhd3eb1b0_0
pandas=1.1.5=py37hf11a4ad_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.3=py37haa95532_1
parso=0.8.1=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=pypi_0
pip=20.3.3=py37haa95532_0
proglog=0.1.9=pypi_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37ha925a31_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.7.3=pyhd3eb1b0_0
pyjwt=1.7.1=py37_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyreadline=2.1=py37_1
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python-engineio=3.5.1=pypi_0
python-socketio=3.1.2=pypi_0
pytz=2020.4=pyhd3eb1b0_0
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyzmq=20.0.0=py37hd77b12b_1
requests=2.25.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.2=py37h9439919_0
seaborn=0.11.0=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=51.0.0=py37haa95532_2
six=1.15.0=py37haa95532_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py37h10aaca4_0
tensorflow-base=2.3.0=eigen_py37h17acbac_0
tensorflow-estimator=2.4.0=pypi_0
tensorflow-gpu=2.3.0=he13fc11_0
tensorflow-gpu-estimator=2.2.0=pypi_0
termcolor=1.1.0=py37_1
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tornado=6.1=py37h2bbff1b_0
tqdm=4.54.1=pypi_0
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.25.11=py_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
yarl=1.6.3=py37h2bbff1b_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
